Added path traversal check

+++ FileSystem.php
@@ -3,22 +3,35 @@
 class FileSystem{
 	
 	public function createFile($name, $contents){
-		$myfile = fopen($name, "w");
-		fwrite($myfile, $contents);
-		fclose($myfile);
-		return "File Created";
+		if($this->checkPath($name)){
+			$myfile = fopen($name, "w");
+			fwrite($myfile, $contents);
+			fclose($myfile);
+			return "File Created";
+		}
+		return "Invalid Name"
 	}
 	
 	public function deleteFile($name){
-		unlink($name);
-		return "File Deleted";
+		if($this->checkPath($name)){
+			unlink($name);
+			return "File Deleted";
+		}
+		return "Invalid Name"
 	}
 	
 	public function viewFile($name){
-		$myfile = fopen($name, "r");
-		$content = fread($myfile, filesize($name));
-		fclose($myfile);
-		return $content;
+		if($this->checkPath($name)){
+			$myfile = fopen($name, "r");
+			$content = fread($myfile, filesize($name));
+			fclose($myfile);
+			return $content;
+		}
+		return "Invalid Name"
+	}
+	
+	public function checkPath($path){
+		return (strpos($path, "..") == false)
 	}
 	
 }
